createAccount with approval

+++ src/main/java/com/asu/secureBankApp/Response/LoginResponse.java
@@ -2,15 +2,15 @@ package com.asu.secureBankApp.Response;
 
 public class LoginResponse {
 
-	int userId;
+	Long userId;
 	
 	Boolean isSuccess;
 	
-	public int getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(int userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
